+++ app/src/frontend/types/course.ts
